backend/requirements.txt
fastapi
uvicorn[standard]
pymysql
cryptography
sqlalchemy
pydantic
python-dotenv
